add shared workspace
